requirements.txt
aiosqlite==0.17.0
aniso8601==7.0.0
annotated-types==0.6.0
anyio==4.2.0
asyncpg==0.29.0
atlastk==0.13.3
bcrypt==4.1.2
click==8.1.7
colorama==0.4.6
fastapi==0.109.0
graphene
graphene-sqlalchemy
graphql-core
graphql-relay
greenlet==3.0.3
h11==0.14.0
idna==3.6
iso8601==1.1.0
libcst==1.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
passlib==1.7.4
promise==2.3
psycopg2
psycopg2-binary
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rich==13.7.0
Rx==1.6.3
setuptools==69.0.3
singledispatch==3.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
uvicorn==0.26.0
